Adding example for non-ideal reactor, folder for py files
This PR has two main changes:
1. Adding a jupyter notebook example to the reactors folder for
a constant-volume, adiabatic reactor.  This is employed within
the context of IDT calculations in high-pressure shock tubes.
I'd be happy to add Comparison to experimental data, pending
publication of a manuscript currently under consideration (and
coauthor permission).

This largely follows the routines established in the batch reactor
example, and states as much in the notebook.

2. It occurs to me that some users might be interested in seeing
the examples as py files, but don't want to bother with installing
Jupyter and figuring out how to use it (while these steps are easy,
they still might represent a barrier to some users).

For the three notebooks in this folder, then, I've saved the notebooks
as simple py files and added them to a folder labeled as such.