Adding Lithium ion battery Jupyter Notebook Example
The new example demonstrates use of the BinarySolutionTabulatedThermo
class, to calculate the open circuit potential for a standard Li ion
battery (graphite anode, LCO cathode) for a range of anode and cathode
Li stoichiometries.

The notebook demonstrates cell potential calculations via two methods:
a kinetic method where the current is set to zero and the correct
electric potentials are fit using scipy.optimize.fsolve, and a
thermodynamic method whereby the equilibrium potential at each
electrode interface is calculated according to the gibbs free energy
of reaction.

The notebook can readily be extended to simulate charge-discharge of
the battery, and preliminary functionality for such an extension is
already programmed into the notebook.